Galaxy/requirements.txt
astroid==4.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
cleo==2.1.0
colorama==0.4.6
connected-components-3d==3.26.0
contourpy==1.3.3
crashtest==0.4.1
cycler==0.12.1
dataclasses==0.6
dcmrtstruct2nii==5
dill==0.4.0
distlib==0.4.0
docx==0.2.4
et_xmlfile==2.0.0
filelock==3.20.0
fonttools==4.60.1
fpdf==1.7.2
fsspec==2025.9.0
idna==3.11
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
itk==5.4.4.post1
itk-core==5.4.4.post1
itk-filtering==5.4.4.post1
itk-io==5.4.4.post1
itk-numerics==5.4.4.post1
itk-registration==5.4.4.post1
itk-segmentation==5.4.4.post1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kmodes==0.12.2
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psutil==7.1.2
psycopg2-binary==2.9.11
pyaml==25.7.0
pydicom==3.0.1
Pygments==2.19.2
pylint==4.0.1
pynrrd==1.1.3
pyparsing==3.2.5
pysera==2.1.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
RapidFuzz==3.14.1
ReliefF==0.1.2
reportlab==4.4.4
requests==2.32.5
rt-utils==1.2.7
scikit-image==0.25.2
scikit-learn==1.7.2
scikit-optimize==0.10.2
scipy==1.16.3
six==1.17.0
sklearn-relief==1.0.0b2
sympy==1.14.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
uml-class-diagram-generator==0.1
urllib3==2.5.0
virtualenv==20.35.3
